fix: rename test files's package name
